+++ src/ResourceManager/Compute/Commands.Compute/Extension/VmssDiskEncryption/GetAzureVmssDiskEncryptionStatus.cs
@@ -28,11 +28,7 @@ using System.Text.RegularExpressions;
 
 namespace Microsoft.Azure.Commands.Compute.Extension.AzureDiskEncryption
 {
-    [Cmdlet(
-        VerbsCommon.Get,
-        ProfileNouns.AzureVmssDiskEncryption),
-        Alias(ProfileNouns.GetAzureRmVmssDiskEncryptionAlias),
-        OutputType(typeof(PSVmssDiskEncryptionStatusContext))]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "VmssDiskEncryption"),Alias(ProfileNouns.GetAzureRmVmssDiskEncryptionAlias),OutputType(typeof(PSVmssDiskEncryptionStatusContext))]
     public class GetAzureVmssDiskEncryptionStatusCommand : VirtualMachineScaleSetExtensionBaseCmdlet
     {
         [Parameter(