+++ src/ResourceManager/Scheduler/Commands.Scheduler/Cmdlets/NewAzureSchedulerServiceBusTopicJobCommand.cs
@@ -26,7 +26,7 @@ namespace Microsoft.Azure.Commands.Scheduler.Cmdlets
     /// <summary>
     /// Creates new service bus topic job.
     /// </summary>
-    [Cmdlet(VerbsCommon.New, "AzureRmSchedulerServiceBusTopicJob", SupportsShouldProcess = true), OutputType(typeof(PSSchedulerJobDefinition))]
+    [Cmdlet("New", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "SchedulerServiceBusTopicJob", SupportsShouldProcess = true), OutputType(typeof(PSSchedulerJobDefinition))]
     public class NewAzureSchedulerServiceBusTopicJobCommand : JobBaseCmdlet, IDynamicParameters
     {
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true, HelpMessage = "The targeted resource group for job.")]