+++ src/ResourceManager/ServiceBus/Commands.ServiceBus/Cmdlets/Queue/RemoveAzureServiceBusQueue.cs
@@ -20,7 +20,7 @@ namespace Microsoft.Azure.Commands.ServiceBus.Commands.Queue
     /// <summary>
     /// 'Remove-AzureRmServiceBusQueue' Cmdlet removes the specified Queue
     /// </summary>
-    [Cmdlet(VerbsCommon.Remove, ServicebusQueueVerb, DefaultParameterSetName = QueuePropertiesParameterSet, SupportsShouldProcess = true), OutputType(typeof(bool))]
+    [Cmdlet("Remove", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "ServiceBusQueue", DefaultParameterSetName = QueuePropertiesParameterSet, SupportsShouldProcess = true), OutputType(typeof(bool))]
     public class RemoveAzureRmServiceBusQueue : AzureServiceBusCmdletBase
     {
         [Parameter(Mandatory = true, ParameterSetName = QueuePropertiesParameterSet, ValueFromPipelineByPropertyName = true, Position = 0, HelpMessage = "The name of the resource group")]
@@ -73,10 +73,17 @@ namespace Microsoft.Azure.Commands.ServiceBus.Commands.Queue
             // delete a Queue 
             if (ShouldProcess(target: Name, action: string.Format(Resources.RemoveQueue, Name, Namespace)))
             {
-                var result = Client.DeleteQueue(ResourceGroupName, Namespace, Name);
-                if (PassThru.IsPresent)
+                try
                 {
-                    WriteObject(result);
+                    var result = Client.DeleteQueue(ResourceGroupName, Namespace, Name);
+                    if (PassThru.IsPresent)
+                    {
+                        WriteObject(result);
+                    }
+                }
+                catch (Management.ServiceBus.Models.ErrorResponseException ex)
+                {
+                    WriteError(ServiceBusClient.WriteErrorforBadrequest(ex));
                 }
             }
         }