+++ src/StackAdmin/Azs.Fabric.Admin/Module/Azs.Fabric.Admin/Generated.PowerShell.Commands/SwaggerPathCommands/Add-AzsScaleUnitNode.ps1
@@ -28,19 +28,23 @@ function Add-AzsScaleUnitNode
     [CmdletBinding()]
     param(
         [Parameter(Mandatory = $true)]
+        [ValidateNotNull()]
         [Microsoft.AzureStack.Management.Fabric.Admin.Models.ScaleOutScaleUnitParameters[]]
         $NodeList,
 
         [Parameter(Mandatory = $false)]
+        [ValidateNotNullOrEmpty()]
         [System.String]
         $ResourceGroupName,
 
         [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [System.String]
         $ScaleUnit,
 
-        [Parameter(Mandatory = $true)]
+        [Parameter(Mandatory = $false)]
         [System.String]
+        [ValidateNotNullOrEmpty()]
         $Location,
 
         [Parameter(Mandatory = $false)]
@@ -82,7 +86,11 @@ function Add-AzsScaleUnitNode
 
     $FabricAdminClient = New-ServiceClient @NewServiceClient_params
 
-    $ParamList = New-ScaleOutScaleUnitParametersListObject -NodeList $NodeList -AwaitStorageConvergence $AwaitStorageConvergence:IsPresent
+    $ParamList = New-ScaleOutScaleUnitParametersListObject -NodeList $NodeList -AwaitStorageConvergence:$AwaitStorageConvergence:IsPresent
+
+    if ([System.String]::IsNullOrEmpty($Location)) {
+        $Location = (Get-AzureRmLocation).Location
+    }
 
     if ([System.String]::IsNullOrEmpty($ResourceGroupName)) {
         $ResourceGroupName = "System.$Location"