+++ src/ResourceManager/ApiManagement/Commands.ApiManagement/ChangeLog.md
@@ -18,9 +18,14 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with default resource groups not being set.
+
+## Version 6.1.3
+* Updated to the latest version of the Azure ClientRuntime.
 
 ## Version 6.1.2
 * Updated help files to include full parameter types and correct input/output types.
+* Updated to the latest version of the Azure ClientRuntime.
 * Fixed issue https://github.com/Azure/azure-powershell/issues/6370
     - Fixed bug in Automapper to translate PsApiManagementApi to ApiContract
 * Fixed issue https://github.com/Azure/azure-powershell/issues/6515