+++ src/ResourceManager/DataLakeAnalytics/Commands.DataLakeAnalytics/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with default resource groups not being set.
+
+## Version 5.1.2
+* Updated to the latest version of the Azure ClientRuntime.
 
 ## Version 5.1.1
 * Updated all help files to include full parameter types and correct input/output types.
@@ -27,7 +31,7 @@
     - Get-AzureRmDataLakeAnalyticsCatalogItemAclEntry
     - Set-AzureRmDataLakeAnalyticsCatalogItemAclEntry
     - Remove-AzureRmDataLakeAnalyticsCatalogItemAclEntry
-* Fixed formatting of OutputType in help files 
+* Fixed formatting of OutputType in help files
 
 ## Version 5.0.0
 * Set minimum dependency of module to PowerShell 5.0
@@ -99,7 +103,7 @@
     - Get-AzureRMDataLakeAnalyticsJobRecurrence
     - Get-AzureRMDataLakeAnalyticsJobPipeline
 * Updated the token audience for job and catalog APIs to use the correct Data Lake specific audience instead of the Azure Resource audience.
-    
+
 ## Version 3.1.0
 
 ## Version 3.0.1