+++ src/ResourceManager/DataMigration/Commands.DataMigration/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with default resource groups not being set.
+
+## Version 0.1.4
+* Updated to the latest version of the Azure ClientRuntime.
 
 ## Version 0.1.3
 * Updated help files to include full parameter types and correct input/output types.