+++ src/ResourceManager/Media/Commands.Media/ChangeLog.md
@@ -18,6 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with default resource groups not being set.
+
+## Version 0.10.2
+* Updated to the latest version of the Azure ClientRuntime.
 
 ## Version 0.10.1
 * Updated help files to include full parameter types and correct input/output types.
@@ -41,7 +45,7 @@
 ## Version 0.8.0
 * Add support for online help
     - Run Get-Help with the -Online parameter to open the online help in your default Internet browser
-    
+
 ## Version 0.7.7
 
 ## Version 0.7.6